@@ -1,8 +1,8 @@
 [tool.poetry]
 name = "clarite"
-version = "2.3.6"
+version = "2.4.0"
 description = "CLeaning to Analysis: Reproducibility-based Interface for Traits and Exposures"
-authors = ["Andre Rico <alr6366@psu.edu>"]
+authors = ["Andre Rico <andreluis.rico@pennmedicine.upenn.edu>"]
 license = "BSD-3-Clause"
 readme = "README.rst"
 repository = "https://github.com/HallLab/clarite-python/"
@@ -16,33 +16,34 @@ classifiers=[
     ]
 
 [tool.poetry.dependencies]
-python = ">=3.8.0,<3.11.0"
+python = ">=3.10,<3.13"
 click = ">7"
-pandas = "^1.3"
+pandas = ">=1.3,<2.2"
 scipy = "^1.9"
 seaborn = ">0.9"
-statsmodels = "^0.13"
+statsmodels = "^0.14"
 matplotlib = "^3.4.2"
-numpy = "^1.24"
-rpy2 = "^3.4.5"
+numpy = ">=1.24,<2.0"
 tzlocal = "^2.1"
-sphinx = {version = "^3.2.1", optional = true}
 numpydoc = {version = "^1.1.0", optional = true}
-sphinx_rtd_theme = {version = "^0.5.0", optional = true}
-sphinx-copybutton = {version = "^0.3.0", optional = true}
 ipython = {version = "^7.18.1", optional = true}
-sphinx-click = {version = "^4", optional = true}
-importlib-metadata = "^5.2.0"
-pandas-genomics = "^0.12"
+pandas-genomics = "^1.0.0"
+rpy2 = "^3.6.0"
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
+mypy = "^1.5.1"
+tox = "^4.26.0"
+pytest = "^7"
 codecov = "^2.1.11"
 coverage = "^5.5"
-flake8 = "^3.9.2"
-pytest = ">=4.6"
+flake8 = "^6"
 pytest-cov = "^2.12.1"
 pytest-xdist = "^2.3.0"
 black = ">=22"
+sphinx = {version = "^4.0.0", optional = true}
+sphinx-click = {version = "^4", optional = true}
+sphinx_rtd_theme = {version = "^0.5.0", optional = true}
+sphinx-copybutton = {version = "^0.3.0", optional = true}
 
 [tool.poetry.extras]
 docs = ["sphinx", "numpydoc", "sphinx_rtd_theme", "sphinx-copybutton", "ipython", "sphinx-click"]
@@ -50,9 +51,9 @@ docs = ["sphinx", "numpydoc", "sphinx_rtd_theme", "sphinx-copybutton", "ipython"
 [tool.poetry.scripts]
 clarite-cli = 'clarite.cli:entry_point'
 
-[tool.poetry.group.dev.dependencies]
-mypy = "^1.5.1"
-
 [build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.tox]
+legacy_tox_ini = true
